Добавлен Api-слой и UserRepository.
